requirements.txt
Package                Version
---------------------- -------
Faker                  16.7.0
mysql-connector-python 8.0.32
pip                    23.0
protobuf               3.20.3
python-dateutil        2.8.2
setuptools             65.6.3
six                    1.16.0
wheel                  0.38.4
